+++ src/extension.ts
@@ -195,6 +195,7 @@ async function activate(context: vscode.ExtensionContext) {
     convertDart2Clj: converters.dart2clj,
     convertJs2Cljs: converters.js2cljs,
     convertHtml2Hiccup: converters.html2hiccup,
+    pasteHtmlAsHiccup: converters.pasteHtmlAsHiccup,
     copyAnnotationHoverText: annotations.copyHoverTextCommand,
     copyJackInCommandToClipboard: jackIn.copyJackInCommandToClipboard,
     copyLastResults: eval.copyLastResultCommand,