+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- [Add commands `calva.convertHtml2Hiccup` and `calva.pasteHtmlAsHiccup`](https://github.com/BetterThanTomorrow/calva/issues/407)
+
 ## [2.0.340] - 2023-03-18
 
 - Workaround: [LiveShare participants incorrectly opening every Clojure file as if via "Open with Notebook"](https://github.com/BetterThanTomorrow/calva/issues/1850)